+++ cs2_visibility/__init__.py
@@ -2,15 +2,31 @@
 
 from .analysis import VisibilityAnalyzer, load_demo_window, export_colored_mesh
 from .models import AnalysisConfig, PlayerPose, VisibilityResult, VisibilityTimelineResult
+from .grenade_lineups import (
+    GrenadeLineup,
+    GrenadeLineupConfig,
+    GrenadeRelease,
+    ThrowPose,
+    ThrowType,
+    derive_grenade_lineup,
+    extract_grenade_lineups,
+)
 from .session import DemoSession
 
 __all__ = [
     "AnalysisConfig",
     "DemoSession",
+    "GrenadeLineup",
+    "GrenadeLineupConfig",
+    "GrenadeRelease",
     "PlayerPose",
+    "ThrowPose",
+    "ThrowType",
     "VisibilityAnalyzer",
     "VisibilityResult",
     "VisibilityTimelineResult",
     "export_colored_mesh",
+    "derive_grenade_lineup",
+    "extract_grenade_lineups",
     "load_demo_window",
 ]